Remove custom implementations of optional and string_view
Summary:
Now that we're using C++17, we don't need to maintain custom implementations of `optional` and `string_view`.

The standard optional has an implicit ctor that causes ambiguities in some reduce_util.h functions. Add explicit overrides to resolve the ambiguities.

Note that this is technically an API break, since exec_aten::optional is changing in a way that can break some existing code. There is no way to roll this out incrementally, though, and there should not be many examples of user code that would break because of this change.

Differential Revision: D64634702